+++ test/acceptance/docker-compose.override.yaml
@@ -1,26 +1,18 @@
-version: "3.7"
+version: '3.8'
 services:
-  wdio-cucumber:
+  ffc-pr-calculator-acceptance-tests:
     environment:
       MAX_INSTANCES: 1
       CHROME_ARGS: "--verbose"
     volumes:
-      - ./features/:/home/node/features/
-      - ./pageobjects/:/home/node/pageobjects/
-      - ./scratch/:/home/node/scratch/
-      - ./steps/:/home/node/steps/
-      - ./support/:/home/node/support/
+      - ./features:/home/node/features
+      - ./pageobjects:/home/node/pageobjects
+      - ./scratch:/home/node/scratch
+      - ./steps:/home/node/steps
+      - ./support:/home/node/support
       - ./babel.config.js:/home/node/babel.config.js
       - ./wdio.conf.js:/home/node/wdio.conf.js
       - ./package.json:/home/node/package.json
-      - ./html-reports/:/home/node/html-reports/
+      - ./html-reports:/home/node/html-reports
     ports:
       - "9228:9229"
-
-  selenium:
-    ports:
-      - "5900:5900"
-
-  hub:
-    ports:
-      - "4444:4444"
