+++ test/acceptance/package.json
@@ -35,7 +35,11 @@
     "@wdio/selenium-standalone-service": "^5.11.0",
     "@wdio/static-server-service": "^5.14.4",
     "axe-webdriverio": "^0.2.0",
+    "browserstack-local": "^1.4.8",
     "chai": "^4.1.2",
+    "cross-env": "^5.2.0",
+    "ecstatic": "^4.1.4",
+    "hosted-git-info": "^4.0.2",
     "selenium-server": "^3.141.59",
     "selenium-standalone": "^6.17.0",
     "wdio-geckodriver-service": "^1.0.0",
@@ -53,7 +57,7 @@
     "chromedriver": "^78.0.1",
     "geckodriver": "^1.18.0",
     "http-server": "^0.12.3",
-    "jest": "^24.8.0",
+    "jest": "^26.6.3",
     "npm-run-all": "^4.1.5",
     "pngjs": "^5.0.0",
     "standard": "12.0.1",
