+++ test/acceptance/wdio.conf.js
@@ -1,38 +1,48 @@
-const { ReportAggregator, HtmlReporter } = require('@rpii/wdio-html-reporter')
+const browserstack = require('browserstack-local')
+const {ReportAggregator, HtmlReporter} = require('@rpii/wdio-html-reporter')
 const log4js = require('@log4js-node/log4js-api')
 const logger = log4js.getLogger('default')
 const envRoot = (process.env.TEST_ENVIRONMENT_ROOT_URL || 'http://host.docker.internal:3000')
-const chromeArgs = process.env.CHROME_ARGS.split(' ')
+const chromeArgs = process.env.CHROME_ARGS ? process.env.CHROME_ARGS.split(' ') : []
 const maxInstances = process.env.MAX_INSTANCES ? Number(process.env.MAX_INSTANCES) : 5
+const user = process.env.BROWSERSTACK_USERNAME
+const key = process.env.BROWSERSTACK_ACCESS_KEY
 
 exports.config = {
-  hostname: 'selenium',
-  path: '/wd/hub',
+  hostname: 'hub-cloud.browserstack.com',
+  user,
+  key,
   specs: ['./features/**/*.feature'],
   exclude: ['./scratch/**'],
+
   maxInstances,
-  capabilities: [{
-    maxInstances,
-    browserName: 'chrome',
-    acceptInsecureCerts: true,
-    'goog:chromeOptions': {
-      args: chromeArgs
+  capabilities: [
+    {
+      maxInstances,
+      pageLoadStrategy: 'none',
+      acceptInsecureCerts: true,
+      browserName: 'chrome',
+      'browserstack.local': true,
+      'browserstack.networkLogs': true,
+      'browserstack.acceptSslCerts': true,
+      'goog:chromeOptions': {
+        args: chromeArgs
+      }
     }
-  }],  
-  //
+  ],
   // ===================
   // Test Configurations
   // ===================
   // Define all options that are relevant for the WebdriverIO instance here
   logLevel: 'warn',
   bail: 0,
-  baseUrl: envRoot + '/selenium:4444',
+  baseUrl: envRoot,
   waitforTimeout: 10000,
   connectionRetryTimeout: 90000,
-  connectionRetryCount: 3,
-  services: ['selenium-standalone'],
+  connectionRetryCount: 1,
+  services: ['browserstack'],
   framework: 'cucumber',
-  specFileRetries: 3,
+  specFileRetries: 0,
   specFileRetriesDelay: 30,
   reporters: ['spec',
     [HtmlReporter, {
@@ -45,7 +55,6 @@ exports.config = {
       LOG: logger
     }]
   ],
-
   // If you are using Cucumber you need to specify the location of your step definitions.
   cucumberOpts: {
     require: ['./steps/**/*.js'], // <string[]> (file/dir) require files before executing features
@@ -63,33 +72,44 @@ exports.config = {
     timeout: 60000, // <number> timeout for step definitions
     ignoreUndefinedDefinitions: false // <boolean> Enable this config to treat undefined definitions as warnings.
   },
-
-  // =====
+  // ====
   // Hooks
   // =====
-  onPrepare: function (config, capabilities) {
+  onPrepare: async function (config, capabilities) {
     const reportAggregator = new ReportAggregator({
       outputDir: './html-reports/',
       filename: 'acceptance-test-suite-report.html',
       reportTitle: 'Acceptance Tests Report',
       browserName: capabilities.browserName
-      // to use the template override option, can point to your own file in the test project:
-      // templateFilename: path.resolve(__dirname, '../template/wdio-html-reporter-alt-template.hbs')
     })
     reportAggregator.clean()
-
     global.reportAggregator = reportAggregator
+    console.log('Connecting local')
+    return new Promise(function (resolve, reject) {
+      exports.bs_local = new browserstack.Local()
+      const bsLocalArgs = {
+        key,
+        verbose: 'true',
+        onlyAutomate: 'true'
+      }
+      exports.bs_local.start(bsLocalArgs, function (error) {
+        if (error) return reject(error)
+        console.log('Connected. Now testing...')
+        resolve()
+      })
+    })
   },
 
   onComplete: function (exitCode, config, capabilities, results) {
     (async () => {
       await global.reportAggregator.createReport()
     })()
+    exports.bs_local.stop()
+    console.log('Testing complete, binary closed')
   },
 
   beforeSession: function () {
     const chai = require('chai')
-
     global.expect = chai.expect
     global.assert = chai.assert
     global.should = chai.should()
@@ -99,14 +119,11 @@ exports.config = {
     if (result.passed) {
       return
     }
-
     const path = require('path')
     const moment = require('moment')
-
     const screenshotFileName = ctx.uri.split('.feature')[0].split('/').slice(-1)[0]
     const timestamp = moment().format('YYYYMMDD-HHmmss.SSS')
     const filepath = path.join('./html-reports/screenshots/', screenshotFileName + '-' + timestamp + '.png')
-
     browser.saveScreenshot(filepath)
     process.emit('test:screenshot', filepath)
   }
